Reborn: Fix signed/unsigned and type-casting issues (3)
Cast mixed integer and float arithmetic to the correct types across the engine, replay tooling, radar, AI, and script logic to avoid newer-compiler warnings and potential runtime issues. The changes keep behavior unchanged while making comparisons and size checks explicit for size_t, UnsignedInt, Real, and legacy frame counters.

+++ Core/GameEngine/Source/Common/INI/INI.cpp
@@ -2421,7 +2421,7 @@ Bool INI::isEndOfBlock( char *bufferToCheck )
 	// it is important to get through all the checks.
 
 	static const char* endString = "End";
-	int endStringLength = strlen(endString);
+	size_t endStringLength = strlen(endString);
 	char restoreChar;
 	char *tempBuff = bufferToCheck;
