+++ Core/GameEngine/Source/Common/System/UnicodeString.cpp
@@ -359,7 +359,7 @@ void UnicodeString::truncateBy(const Int charCount)
 		{
 			ensureUniqueBufferOfSize(len + 1, true, nullptr, nullptr);
 			size_t count = charCount;
-			if (charCount > len)
+			if ((size_t)charCount > len)
 			{
 				count = len;
 			}
@@ -376,7 +376,7 @@ void UnicodeString::truncateTo(const Int maxLength)
 	if (m_data)
 	{
 		const size_t len = wcslen(peek());
-		if (len > maxLength)
+		if (len > (size_t)maxLength)
 		{
 			ensureUniqueBufferOfSize(len + 1, true, nullptr, nullptr);
 			peek()[maxLength] = 0;
